Refactor detection scripts: unify BoundingBox model, eliminate dict soup and temp files
- Single BoundingBox source of truth in models/product.py (int -> float)
- Gemini functions return typed Pydantic models instead of dicts
- Gemini functions accept image_data bytes instead of file paths
- image.py functions accept typed BoundingBox and PIL Image directly
- Pipeline uses model attributes instead of dict key access throughout
- Removed: unused imports, dead variables, inline imports, temp file dance

+++ server/app/models/product.py
@@ -5,10 +5,10 @@ from pydantic import BaseModel, Field
 
 
 class BoundingBox(BaseModel):
-    x: int
-    y: int
-    width: int
-    height: int
+    x: float
+    y: float
+    width: float
+    height: float
 
 
 class ProductData(BaseModel):
